Move send-access components into their own folder
