+++ apps/web/src/app/oss.module.ts
@@ -4,7 +4,7 @@ import { AuthModule } from "./auth";
 import { LoginModule } from "./auth/login/login.module";
 import { TrialInitiationModule } from "./billing/trial-initiation/trial-initiation.module";
 import { LooseComponentsModule, SharedModule } from "./shared";
-import { AccessComponent } from "./tools/send/access.component";
+import { AccessComponent } from "./tools/send/send-access/access.component";
 import { OrganizationBadgeModule } from "./vault/individual-vault/organization-badge/organization-badge.module";
 import { VaultFilterModule } from "./vault/individual-vault/vault-filter/vault-filter.module";
