+++ apps/web/src/app/tools/send/send-access/index.ts
@@ -1,0 +1,2 @@
+export { AccessComponent } from "./access.component";
+export { SendAccessExplainerComponent } from "./send-access-explainer.component";
